Minor Updates
Added a place holder for foriegn objects in the wind level and updated
the menu so that we can go to any level form there.  Also, I don't know
what happened, but the tank level disappeared.  There wasn't much on
there anyway though. That being said, the button for it on the menu is
there, but it doesn't work